+++ backend/run.py
@@ -1,4 +1,4 @@
-from application import create_app
+from app import create_app
 
 app = create_app()
